+++ jaehoon/java-mid1/src/main/java/section06/BankAccount.java
@@ -6,22 +6,22 @@ public class BankAccount {
 
   private final String accountNumber;
 
-  private final double balance;
+  private final long balance;
 
   // NOTE: accountNumber는 필수값이며, balance는 0 이상의 값이 들어온다고 가정함 (입력값을 받을 때 체크함)
-  public BankAccount(String accountNumber, double balance) {
+  public BankAccount(String accountNumber, long balance) {
     this.accountNumber = accountNumber;
     this.balance = balance;
   }
 
-  public BankAccount deposit(double amount) {
+  public BankAccount deposit(long amount) {
     if (amount <= 0) {
       throw new IllegalArgumentException("0보다 큰 금액을 입금해야 합니다.");
     }
     return new BankAccount(accountNumber, (balance + amount));
   }
 
-  public BankAccount withdraw(double amount) {
+  public BankAccount withdraw(long amount) {
     if (amount <= 0) {
       throw new IllegalArgumentException("0보다 큰 금액을 출금해야 합니다.");
     }
